+++ packages/terra-date-time-picker/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Screenshots updated for mobile view.
+* Update wdio test to work with changes in terra-date-picker
 
 ### Changed
 * Renamed test and example SCSS files to end with `.module.scss` extension